+++ ParallaxGenLib/src/patchers/PatcherShaderTransform.cpp
@@ -1,7 +1,28 @@
 #include <utility>
 
+#include "NIFUtil.hpp"
+#include "ParallaxGenUtil.hpp"
 #include "patchers/PatcherShaderTransform.hpp"
 
+#include <spdlog/spdlog.h>
+
 PatcherShaderTransform::PatcherShaderTransform(std::filesystem::path NIFPath, nifly::NifFile *NIF,
-                                               std::string PatcherName)
-    : Patcher(std::move(NIFPath), NIF, std::move(PatcherName)) {}
+                                               std::string PatcherName, const NIFUtil::ShapeShader &From,
+                                               const NIFUtil::ShapeShader &To)
+    : Patcher(std::move(NIFPath), NIF, std::move(PatcherName)), FromShader(From), ToShader(To) {}
+
+std::unordered_set<std::tuple<std::filesystem::path, NIFUtil::ShapeShader, NIFUtil::ShapeShader>,
+                   PatcherShaderTransform::ErrorTrackerHasher>
+    PatcherShaderTransform::ErrorTracker;
+
+void PatcherShaderTransform::postError(const std::filesystem::path &File) {
+  if (ErrorTracker.insert({File, FromShader, ToShader}).second) {
+    spdlog::error(L"Failed to transform from {} to {} for {}",
+                  ParallaxGenUtil::ASCIItoUTF16(NIFUtil::getStrFromShader(FromShader)),
+                  ParallaxGenUtil::ASCIItoUTF16(NIFUtil::getStrFromShader(ToShader)), File.wstring());
+  }
+}
+
+auto PatcherShaderTransform::alreadyTried(const std::filesystem::path &File) -> bool {
+  return ErrorTracker.contains({File, FromShader, ToShader});
+}